requirement.txt
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
Django==2.2
django-ckeditor==6.0.0
django-haystack==3.0
django-js-asset==1.2.2
django-redis==4.12.1
django-simpleui==2021.4.1
idna==2.10
itsdangerous==1.1.0
kombu==5.0.2
lxml==4.6.3
Pillow==8.1.2
prompt-toolkit==3.0.18
py3Fdfs==2.2.0
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Whoosh==2.7.4
